P-43 Refactor comment search result
