+++ frontend/src/features/CommentWidget/index.ts
@@ -1,0 +1,1 @@
+export { CommentWidget } from "./CommentWidget";
