+++ frontend/src/features/CommentsSerchResults/CommentsSearchResults.tsx
@@ -1,20 +1,8 @@
-import { CommentDto } from "@/entities/comments/api/dto/comment.dto";
-import { Card, CardContent, Typography } from "@mui/material";
+import { CommentWidget } from "../CommentWidget";
+import { Comment } from "@/entities/comments/model/Comment";
 
-export const CommentsSearchResults = (props: { results: CommentDto[]; }) => {
-    return (props.results.map((comment: CommentDto, index: number) => (
-        <Card key={index} sx={{ width: "100%" }}>
-            <CardContent>
-                <Typography variant="body1">
-                    <strong>Пользователь:</strong> {comment.from.username}
-                </Typography>
-                <Typography variant="body2" color="text.secondary">
-                    {new Date(comment.date).toLocaleString()}
-                </Typography>
-                <Typography variant="body2" sx={{ marginTop: 1 }}>
-                    {comment.message}
-                </Typography>
-            </CardContent>
-        </Card>
+export const CommentsSearchResults = (props: { results: Comment[]; }) => {
+    return (props.results.map((comment: Comment) => (
+        <CommentWidget key={comment.tlgId} {...comment} />
     )));
 };
